feat: create guest session id

+++ src/app/services/favoritesService.ts
@@ -1,31 +1,46 @@
 "use client";
 import axios from 'axios';
 
-export const addFavorite = async (movieId: string) => {
+export const createGuestSession = async () => {
   try {
-    const response = await axios.post('/api/movies/favorites/add-favorite', {
-      movieId,
-    });
-    console.log(response.data.message);
+    const response = await axios.get('/api/guest-session');
+    return response.data;
+
   } catch (error) {
-    console.error('Error adding movie to favorites:', error);
+    console.error('Failed to create guest session >>>:', error);
+    throw error;
   }
 };
 
-export const removeFavorite = async (movieId: string) => {
+export const addFavorite = async (guestSessionId: string, movieId: string) => {
   try {
-    const response = await axios.delete('/api/movies/favorites/remove-favorite', {
-      data: { movieId }
+    const response = await axios.post('/api/movies/favorites/add-favorite', {
+      guestSessionId,
+      movieId,
     });
-    console.log(response.data.message);
+  } catch (error) {
+    console.error('Error adding movie to favorites>>>:', error);
+  }
+};
+
+export const removeFavorite = async (guestSessionId: string, movieId: string) => {
+  try {
+    if (!guestSessionId || !movieId) {
+      throw new Error('Guest session ID and movie ID are required');
+    }
+    const response = await axios.delete('/api/movies/favorites/remove-favorite', {
+      data: { guestSessionId, movieId }
+    });
   } catch (error) {
     console.error('Error removing movie from favorites:', error);
   }
 };
 
-export const getFavorites = async (): Promise<string[]> => {
+export const getFavorites = async (guestSessionId: string): Promise<any[]> => {
   try {
-    const response = await axios.get('/api/movies/favorites/get-favorite');
+    const response = await axios.get('/api/movies/favorites/get-favorite', {
+      params: { guestSessionId },
+    });
     return response.data;
   } catch (error) {
     console.error('Error fetching favorite movies:', error);
